Checking users for authorization and matching IDs in endpoints

+++ src/app/api/comments/delete/route.js
@@ -1,30 +1,52 @@
-    import { deleteCommentById } from '../../../../lib/db'
-    import { NextResponse } from 'next/server'
-    import { idParamSchema } from '../../../../lib/zod'
-
-    export async function DELETE(request) {
-        try {
-            const { searchParams } = new URL(request.url)
-            const validation = idParamSchema.safeParse({
-                id: searchParams.get('id')
-            })
+import { deleteCommentById, fetchCommentFromDBById } from '../../../../lib/db'
+import { NextResponse } from 'next/server'
+import { idParamSchema } from '../../../../lib/zod'
+import { authConfig } from "@/lib/auth"
+import { getServerSession } from "next-auth/next"
 
 
-            if (!validation.success) {
+export async function DELETE(request) {
 
-                console.log("backend Zod error:", validation.error.format())
-                return NextResponse.json({ error: "Invalid data format" }, { status: 400 })
-            }
-            const { id } = validation.data
-            await deleteCommentById(id)
-            return NextResponse.json({
-                message: "comments deleted successfully",
-                success: true,
+    const session = await getServerSession(authConfig)
+    try {
 
-            }, { status: 200 })
-        } catch (error) {
-            return NextResponse.json({ error: "Internal Server Error", details: error.message }, { status: 500 })
+
+
+        if (!session.user) {
+            return NextResponse.json({ error: "The user is not authorized" }, { status: 401 })
         }
 
 
+
+        const { searchParams } = new URL(request.url)
+        const validation = idParamSchema.safeParse({
+            id: searchParams.get('id')
+        })
+
+
+        if (!validation.success) {
+
+            console.log("backend Zod error:", validation.error.format())
+            return NextResponse.json({ error: "Invalid data format" }, { status: 400 })
+        }
+        const { id } = validation.data
+
+        const currentComment = await fetchCommentFromDBById(id)
+
+        if (currentComment.user_id !== session.user.id) {
+            return NextResponse.json({ error: "wrong user" }, { status: 403 })
+        }
+
+        await deleteCommentById(id)
+        return NextResponse.json({
+            message: "comments deleted successfully",
+            success: true,
+
+        }, { status: 200 })
+    } catch (error) {
+
+        return NextResponse.json({ error: "Internal Server Error", details: error.message }, { status: 500 })
     }
+
+
+}
